Next iteration of the code
Now, we have data on GB and SPF forecasts, as well as observed values of the corresponding variables that were forecasted taken from FRED using a very helpful set of documentation (https://www.briancjenkins.com/fredpy/docs/build/html/fredpy_examples.html).

Next steps are to remove some rows from the GB forecasts so that all three dataframes align with the required data. Then, we combine the data and perform calculations for the forecasted and observed annual changes in the variables